selenium_based/requirements.txt
requests
PyYAML
selenium
selenium-wire
undetected-chromedriver
chromedriver-autoinstaller
selenium-stealth
tqdm
beautifulsoup4
fake-useragent